Логика подсчета через int заменена на double

+++ src/main/shapes/Circle.java
@@ -1,7 +1,5 @@
 package main.shapes;
 
-import main.workers.WorkerInfo;
-
 /**
  * @author Pstyga Stas
  * @created 19.03.2026
@@ -15,12 +13,12 @@ public class Circle implements Shape {
     }
 
     @Override
-    public int getSquare() {
-        return (int) Math.round(Math.PI * Math.pow(this.radius, 2));
+    public double getSquare() {
+        return Math.round(Math.PI * Math.pow(this.radius, 2));
     }
 
     @Override
-    public int getPerimeter() {
-        return (int) Math.round(2 * Math.PI * this.radius);
+    public double getPerimeter() {
+        return Math.round(2 * Math.PI * this.radius);
     }
 }
