+++ src/main/shapes/Square.java
@@ -4,7 +4,7 @@ package main.shapes;
  * @author Pstyga Stas
  * @created 19.03.2026
  */
-public class Square implements Shape{
+public class Square implements Shape {
 
     private int sideA;
     private int sideB;
@@ -15,12 +15,12 @@ public class Square implements Shape{
     }
 
     @Override
-    public int getSquare() {
+    public double getSquare() {
         return this.sideA * this.sideB;
     }
 
     @Override
-    public int getPerimeter() {
+    public double getPerimeter() {
         return (this.sideA + this.sideB) * 2;
     }
 }
